Add new GLM test feature
Add new statistical test - instead of wilcoxon test, fit a glm negative binomial model to the data with a null model, then one with the proposed model. Then compare the models to find a p-value - as would be performed in differential expression testing. New parameter into the function for specifying whteher wilcoxon or glm test should be performed.